requirements.txt
Flask==3.1.2
flask-restx==1.3.2
flask-cors==6.0.2
gunicorn==21.2.0
requests==2.32.5

azure-core==1.37.0
azure-storage-blob==12.27.1
msal==1.34.0

cryptography==46.0.3
PyJWT==2.10.1
